process_services_layer/telegram_wrapper/requirements.txt
flask==1.1.1
python-telegram-bot==12.2.0
